Added vimhalla first version.

+++ install.sh
@@ -25,22 +25,22 @@ if test -e $HOME/.shell_prompt_no_pl.sh -a \! -e $HOME/.shell_prompt_no_pl_bkg.s
   mv $HOME/.shell_prompt.sh $HOME/.shell_prompt_bkg.sh || return $mvzshrcerrno;
 fi
 
-#if test -e $HOME/.vimrc -a \! -e $HOME/.vimrc_bkg; then
-#  mv $HOME/.vimrc $HOME/.vimrc_bkg || return $mvvimrcerno;
-#fi
-#
-#if test -e $HOME/.vim \! -e $HOME/.vim_bkg; then
-#  mv $HOME/.vim $HOME/.vim_bkg || return mvvimerno;
-#fi
+if test -e $HOME/.vimrc -a \! -e $HOME/.vimrc_bkg; then
+  mv $HOME/.vimrc $HOME/.vimrc_bkg || return $mvvimrcerno;
+fi
+
+if test -e $HOME/.vim -a \! -e $HOME/.vim_bkg; then
+  mv $HOME/.vim $HOME/.vim_bkg || return $mvvimerno;
+fi
 
 ln -s $HOME/DotFiles/zshrc $HOME/.zshrc
 ln -s $HOME/DotFiles/dircolors.256dark $HOME/.dircolors.256dark
 ln -s $HOME/DotFiles/shell_prompt.sh $HOME/.shell_prompt.sh
 ln -s $HOME/DotFiles/shell_prompt_no_pl.sh $HOME/.shell_prompt_no_pl.sh
 
-#ln -s $HOME/DotFiles/vimrc $HOME/.vimrc
+ln -s $HOME/DotFiles/vimrc $HOME/.vimrc
 
-#ln -s $HOME/DotFiles/vim $HOME/.vim
-#git clone https://github.com/gmarik/Vundle.vim.git $HOME/DotFiles/vim/bundle/Vundle.vim
-#git clone 
+ln -s $HOME/DotFiles/vim $HOME/.vim
+git clone https://github.com/gmarik/Vundle.vim.git $HOME/DotFiles/vim/bundle/Vundle.vim
+vim \+BundleInstall\! \+BundleClean \+qall
